move to utils dir
